Fix missing sidebar in Chrome 34 and later
See #29.

I updated blink-devtools to the latest (requires sebmarkbage/blink-devtools#1) and did a manual merge, changing ReactPanel.js to match all changes that happened to ElementPanel.js since the last devtools version.

I'm not sure we actually want to continue tracking the latest devtools; we could presumably keep an old version of the panels and just fix the specific problem with the removed SidebarView.

This doesn't actually work properly yet due to some style-related bug; when the panel is rendered, the View components read all of their dimensions as 0, then 0 is set as the actual width and height of each element, causing nothing to display. However, if the script debugger is used within `ReactPanel.show` to inspect `this.element` (but not to change any properties on it), then the panel seems to display correctly. I'm not sure how to go about fixing this. I can look more later, but if anyone has ideas here I'd appreciate it.