Improve rendering perf when expanding large objects in sidebar panel.
If you have an object with many properties, particularly those with getters and setters, expanding the object tree in the right sidebar panel in the React tab can take upwards of 10 seconds or more.

This patch adds an overflow: auto; rule to any elements with class="fill" which prevents Chrome from rendering content outside of its bounding box and vastly improves performance for large object expansions.