A bunch of hacks to support 0.13
In 0.13 we no longer have access to ReactComponentMixin so we'll monkey
patch the ReactReconciler module instead.

Also, fixed a quirk around text components so that it is now possible
to see and edit text in the middle of a sequence of components.

Ultimately, we'll need a better long term strategy for this.