:bug: Allow you to expand/collapse tree elements
In newer versions of Chrome you can't expand any tree element (e.g. I
can't expand/collpse the state props).

The cause was this line of code which I'm guessing is using a
depreciated API:
https://github.com/sebmarkbage/blink-devtools/blob/master/Source/devtools/front_end/treeoutline.js#L943.

I'm guessing the correct thing to do would be to update blink-devtools
to the latest however as a stop gap we can update
TreeElement.prototype.isEventWithinDisclosureTriangle with its latest
implementation:
https://chromium.googlesource.com/chromium/blink/+/master/Source/devtools/front_end/ui/treeoutline.js.
It's far from ideal but it solves the problem.

Fixes #63.