Fix webpack.config to exclude node_modules
If it's a string it actually has to be an absolute path to exclude. Using a
regex like this seems to be the norm that most people do.

The benefit is that we no longer transform code like React's which is already
transformed.

We also don't use `.jsx` files, so I dropped that support from the regex too.

The `helper.js` file doesn't seem to exist anymore, so I removed it from the
exclusion list.