Stringify element references (kind of)
This will just stringify an element reference to document.create(‘HTMLDivElement’).
Ideally this would be document.create(‘div’) but since there’s no benefit from making it happen I’ll avoid it for now.

Treating element references as objects was a source of issues because the React bridge isn’t able to inspect the properties of some element’s prototypes.
E.g.: HTMLDivElement.prototype['align’], Chrome fails to evaluate this expression with an “Illegal invocation”. As far as I understood align is a getter but is passed the wrong this to it.