Fix #248, when an <option> contains raw strings
The bug is easily replicable (see
https://github.com/facebook/react-devtools/issues/248), and the problem is
that `element` is just a string in that case, such that there is no
`element._currentElement`.
This also adds a check to see if element is an object, and if not, assumes
that it is a raw value that will be displayed as a string.

Test Plan:
Go to a page with `<option>some text {1}</option>`, see that devtools doesn't
break.