Add frontend for highlight all option
Added a plugin folder for the colorizer to live in and a colorizer checkbox to settings pane. I moved the event wrapper from the settings pane into each individual checkbox, so interaction wouldn't toggle both options. I also added event broadcasters where necessary for change in colorizer selection.

Now placeholder text is passed in to SearchPane from the store. Whenever a change in state is registered on the highlight checkbox, it updates the search with the new text. I also added a constant for default placeholder text.

I moved ControlState to the general types file in frontend. I also made sure the backend expects this type as well. I had to use an invariant on toJS() because flow was complaining for some reason.