Add ability to highlight all searched items
When searching, if highlight box is selected all valid items are highlighted. I changed the style of the MultiOverlay so the overlays will scroll.

I also made separate functions for highlighting, so calls to the bridge are only done in one place. It made it easier to handle logic for when to highlight and when not to.