Add option to search using Regular Expressions
There's now a checkbox in the settingsPane to allow searching by regular expressions. nodeMatchesText has a function with the option to use regex to check for valid components.

I added a state to the store that force reloads the search. This is needed for when the regex option is changed. Previously the search would not reload because the string itself hasn't changed. Also, when regex is enabled the search has to start from scratch, unlike for normal text where it can be filtered down.