Fix _mountComponent for use after after initial rendering
When a component was mounted after the initial rendering of the page, it was being added to the searchRoots before its parent was set. This caused errors with getting the parent id from nodeMatchesText.

Now, the check to see if it fits the current search occurs in _updateComponent, after the parent has been set. The highlighting is also rerendered after any valid component update or component unmounting.