Add refresh for multiOverlay when scroll occurs
Whenever scrolling occurs for any element on the page, the multiOverlay is redrawn to match the new element positions. This fixes the updating problem for when certain elements on a page are fixed while others are not.

The redrawing is throttled, so the new positions are calculated only when the page is redrawn. This should prevent any scrolling performance issues. All of this occurs on the backend, so no notifications are passed through the bridge about redraw updates.

I also made the ControlState type a union with Record type, so the use of an invariant isn't needed anymore.