Optimize redrawing and clean up scrolling updates
I removed one of the functions for scroll updates that made it unnecessarily complex.

Also, now element overlays are only drawn if they are visible in the document. This is calculated based on current scroll position. Some minimal slowdown is still possible if there is a large number of overlays, but it is basicaly eliminated with this.