Devtools for everyone!!!
Just add `<script src="http://localhost:8097"></script>` to the top of any
html file, and magically debug the react running there. (swap out your
network-local IP for `localhost` for pages not on your machine)

This works for:
- iframes
- safari
- mobile safari/chrome/whatever
- embedded webviews
- like anything

Test Plan:
- `npm run build` in `/shells/electron`
- then `npm start`

Add the script tag to the top of an html file.
Load the html file anywhere.
Debug to your heart's content.