Remove 'null' and other invalid children when we set 'children'
**what is the change?:**
In the place where we get the 'children' for an element with React
'stack', we want to filter out 'null' and other invalid children. In at
least one case it causes an error later because we expect only valid
children.

**why make this change?:**
To fix a bug.
We previously tried ignoring 'null' children in a place closer to where
the error is thrown, but this makes more sense and will avoid unexpected
situations in other places in the code.

**test plan:**
Manually tested;

1. Made CRA which renders the following:
```
class App extends Component {
  render() {
    const worldServers = [
      {
        name: 'foo',
        delay: null,
      },
      {
        name: null,
        delay: 'bar',
      },
      {
        name: null,
        delay: null,
      },
    ];
    let serverList = []
    for (let i in worldServers) {
        serverList.push(
            <option key={i} value={i}>{worldServers[i].name}{worldServers[i].delay}</option>
        )
    }
    return (
      <div className="App">
        <div className="App-header">
          <img src={logo} className="App-logo" alt="logo" />
          <h2>Welcome to React</h2>
        </div>
        <p className="App-intro">
          To get started, edit <code>src/App.js</code> and save to reload.
        </p>
        <select>{serverList}</select>
      </div>
    );
  }
}
```
2.  Run the CRA app and visit that url when testing the React devtools
using yarn test:chrome
3.  See that with this fix it actually renders the component tree, and
without it the React devtools won't load.