Copy 'traverseAllChildrenImpl' from React and use in 'getData'
**what is the change?:**
Rather than writing our own logic for filtering the children in React
for the stack-specific code that gets the data to be rendered in the
React devtools, it seems more reliable to get an exact copy of the logic
that React uses to traverse the children in 'React.Children'.
The 'traverseAllChildrenImpl' and helper functions in that file are
copied from https://github.com/facebook/react/blob/240b84ed8e1db715d759afaae85033718a0b24e1/src/isomorphic/children/ReactChildren.js

**why make this change?:**
We want to ignore 'null' children and when we get the raw 'children'
object it doesn't have a guaranteed type or iterator.

**test plan:**
Manually tested - Flarnie will add a screenshot in the comments.

**issue:**
fixes #845