Prevent error if global hook couldn't run
On pages with a CSP that prevents loading of scripts at self the global
hook will fail. This is because we create a script tag and then inject
it into the page itself.

When this happens window.__REACT_DEVTOOLS_GLOBAL_HOOK__ will be
undefined, and our checks for react will throw an exception.