Fully embrace yarn and its workspaces (#978)
Given the `yarn.lock` in the root of this repo, tools like Travis and
users are likely to manage dependencies with yarn. However, many of the
`scripts` call out to `npm`, resulting in mixing package managers. The
version of lerna here at the time was also coupled to npm, and it now
supports yarn.

This moves nearly all references from npm to yarn in package.json's `scripts`,
build scripts, READMEs, travis configuration, and lerna configuration.

This enables yarn's workspaces with the `packages/` directory. It also updates
lerna to its latest version which supports yarn and configures it to use workspaces.

Lerna is now only really needed for its bulk versioning and publishing functionality,
so I removed it from the `postinstall` step [(which leaving it in incidentally also
causes it to stall
indefinitely)](https://github.com/lerna/lerna/issues/1001)

This also adds a `.yarnrc` forcing the yarn client to run post-install
scripts for dependencies. The `electron` dependency for the standalone
devtools relies on this, and while it is the default for yarn, more
secure configurations disable it.

Test plan:
  * `yarn lint && yarn typecheck && yarn test`
  * `yarn build:extension && yarn test:chrome`
    * I ran the test for the firefox extension as well but doesn't seem to pick up
      that sites are using react and doesn't activate. This happens on
      the master branch as well.
  * `yarn build:standalone && yarn test:standalone`
  * Let travis run and pass

I'm unsure how to test lerna's publishing functionality. Hopefully
someone can help with that part :)