Remove Relay classic plugin and any reference to Relay
With the deprecation, of Relay Classic it no longer makes sense to maintain this plugin.

Test strategy
I follow the existing instructions for test in Chrome, Firefox and standalone, making sure a few smoke tests passed when using code that still uses Relay classic and React. Below are some screenshots from the console.