Stop sending profile data when recording is off
In https://github.com/facebook/react/issues/14574, I have reported an issue where the devtools can get really slow.
It happens when a lot of profiling data are sent via the bridge to the frontend.

This PR disable sending the profiling data to the frontend when the profiler isn't recording.

I wasn't sure where to prevent the data from being sent and how to do it. Also, I'm not sure if not sending this data when not recording can break other features.