Prevent invocation of getters/setters during Object inspection
Previously, the code was using val.__proto__[name] to retrieve prototype values.
Unfortunately, when dealing with getter and setters, this invokes the
function itself, instead of referring to it. This can have side-effects,
including crashes.
By using the property descriptors, we have clean access to the prototype
values. Note that the code could be expanded to provide on-demand
getter inspections similar to the Chrome devtools inspections.

Note: it would be preferable to use getOwnPropertyDescriptors() instead
of getOwnPropertyNames() + getOwnPropertyDescriptor(), but flowcheck
does not have the getOwnPropertyDescriptors function registered.